persist-types: extract low-level persist types for better dep hygiene
The repr crate, which is supposed to be pretty low-level, had to depend on the
entirety of the persist crate (which is the entire implementation) and all its
sub-deps just for one trait definition. I noticed that aljoscha's upcoming
source work was going to add a second instance of this for the dataflow-types
crate, which is similarly low-level like repr. This has been bothering me for a
bit, so seems like a good time to fix it.
