+++ src/persist-types/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "persist-types"
+description = "Types for the persist crate."
+version = "0.0.0"
+edition = "2018"
+publish = false
+
+# NB: This is meant to be a strong, independent abstraction boundary, please
+# don't leak in deps on other Materialize packages.
+[dependencies]
