+++ src/persist/Cargo.toml
@@ -28,6 +28,7 @@ futures-util = "0.3.16"
 futures-executor = "0.3.16"
 log = "0.4.13"
 ore = { path = "../ore", default-features = false, features = ["metrics"] }
+persist-types = { path = "../persist-types" }
 timely = { git = "https://github.com/TimelyDataflow/timely-dataflow", default-features = false, features = ["bincode"] }
 tokio = { version = "1.11.0", default-features = false, features = ["macros", "sync"] }
